requirements.txt
Django
psycopg2-binary
ipython
pandas
numpy
folium
django-polymorphic
django-money
djangorestframework
markdown
django-filter
djangorestframework-gis
